requirements.txt
pandas
numpy
notebook
pyYAML
mlflow
seaborn
scikit-learn
dvc
-e .